Bug fix versions of Mesquite 1.1 (1.1.0, 1.1.1, 1.1.2) should have been tags in the 1.1 branch, rather than a separate branch in the CVS repositry. Better late than never... merge changes from 1.1.1 branch back into 1.1 branch.
Merge fixes from CVS HEAD into 1.1 branch:
  - Mike's fix for configure script on HP
  - Mike's fix for MeshWriter on 64-bit platforms
  - Update of files in config_aux/ for building on Mac

Update version string in MesquiteVersion.cpp to 1.1.3

Tag as MESQUITE_1_1_3 in CVS.

+++ src/Mesh/MeshWriter.cpp
@@ -162,7 +162,7 @@ void EdgeIterator::get_adjacent_vertices( MsqError& err )
       // For each edge
     for (size_t d = 0; d < num_edges; ++d)
     {
-      const size_t* edge = TopologyInfo::edge_vertices( type, d, err );
+      const unsigned* edge = TopologyInfo::edge_vertices( type, d, err );
       MSQ_ERRRTN(err);
       size_t vert1 = elem.get_vertex_index( edge[0] );
       size_t vert2 = elem.get_vertex_index( edge[1] );
